+++ src/functions/models/odata-query.ts
@@ -21,35 +21,6 @@ export interface ODataQuery<T> {
   select<Tkey extends keyof T>(...keys: SelectParams<T, Tkey>): ODataQuery<T>
 
   /**
-   * Adds a $orderby operator to the OData query.
-   * Ordering over relations is supported (check OData implementation for details).
-   *
-   * @param key key in T.
-   * @param order the order of the sort.
-   *
-   * @example
-   *
-   * q.orderBy('blogs', 'desc')
-   */
-  orderBy<TKey extends keyof T>(
-    key: TKey,
-    order?: 'asc' | 'desc'
-  ): ODataQuery<T>
-  /**
-   * Adds a $orderby operator to the OData query.
-   * Ordering over relations is supported (check OData implementation for details).
-   *
-   * @param exp a lambda expression that builds the orderby expression from the builder.
-   *
-   * @example
-   *
-   * q.orderBy(u => u.blogs().id.desc())
-   */
-  orderBy(
-    exp: (ob: OrderByBuilder<T>) => OrderBy | OrderByExpression
-  ): ODataQuery<T>
-
-  /**
    * Adds a $filter operator to the OData query.
    * Multiple calls to Filter will be merged with `and`.
    *
@@ -75,6 +46,60 @@ export interface ODataQuery<T> {
   ): ODataQuery<T>
 
   /**
+   * Adds a $orderby operator to the OData query.
+   * Ordering over relations is supported (check OData implementation for details).
+   *
+   * @param key key in T.
+   * @param order the order of the sort.
+   *
+   * @example
+   * q.orderBy('blogs', 'desc')
+   */
+  orderBy<TKey extends keyof T>(
+    key: TKey,
+    order?: 'asc' | 'desc'
+  ): ODataQuery<T>
+  /**
+   * Adds a $orderby operator to the OData query.
+   * Ordering over relations is supported (check OData implementation for details).
+   *
+   * @param exp a lambda expression that builds the orderby expression from the builder.
+   *
+   * @example
+   * q.orderBy(u => u.blogs().id.desc())
+   */
+  orderBy(
+    exp: (ob: OrderByBuilder<T>) => OrderBy | OrderByExpression
+  ): ODataQuery<T>
+
+  /**
+   * Adds a $skip and $top to the OData query.
+   * The pageindex in zero-based.
+   * This method automatically adds $count=true to the query.
+   *
+   * @param pagesize page index ($skip).
+   * @param page page size ($top)
+   *
+   * @example
+   * q.paginate(50, 0)
+   */
+  paginate(pagesize: number, page?: number): ODataQuery<T>
+  /**
+   * Adds a $skip and $top to the OData query.
+   * The pageindex in zero-based.
+   *
+   * @param options paginate query options
+   *
+   * @example
+   * q.paginate({ pagesize: 50, page: 0, count: false })
+   */
+  paginate(options: {
+    pagesize: number
+    page?: number
+    count?: boolean
+  }): ODataQuery<T>
+
+  /**
    * Adds a $expand operator to the OData query.
    * Multiple calls to Expand will expand all the relations, e.g.: $expand=rel1(...),rel2(...).
    * The lambda in the second parameter allows you to build a complex inner query.
@@ -83,7 +108,6 @@ export interface ODataQuery<T> {
    * @param query a lambda expression that build the subquery from the querybuilder.
    *
    * @example
-   *
    * q.exand('blogs', q => q.select('id', 'title'))
    */
   expand<key extends keyof RelationsOf<T>, U = T[key]>(
@@ -95,7 +119,6 @@ export interface ODataQuery<T> {
    * exports query to string joined with `&`
    *
    * @example
-   *
    * '$filter=order gt 5&$select=id'
    */
   toString(): string
@@ -104,7 +127,6 @@ export interface ODataQuery<T> {
    * exports query to object key/value
    *
    * @example
-   *
    * {
    *  '$filter': 'order gt 5',
    *  '$select': 'id'
